update db package to use the new drizzle beta features including RQB V2
